config parser now can go back and forth between python and XML
